add doc on exported C types
This adds documentation to the exported types and provides a cbindgen
configuration for `snips-nlu-ontology-ffi` and
`snips-nlu-ontology-ffi-with-parsers` crates.

The cbingen configuration uses the macro expand feature of cbindgen,
make sure you use a recent enough cbindgen an switch to the nightly
compiler before generating bindings.

This also updates the version of ffi-utils for one with documentation on
public types